requirements.txt
python-dotenv
chromadb
langchain-text-splitters
google-genai
google-generativeai
openai-whisper
edge-tts
sounddevice
scipy
numpy
pygame
streamlit
